Files separated & Directory structure updated
